--- src/components/pages/labs/pagination/index.js
@@ -1,3 +1,0 @@
-import Pagination from './pagination';
-
-export default Pagination;
